Move set_default and fix lint/types
